pyproject.toml
[tool.poetry]
name = "mashiropy"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
psutil = "^5.8.0"
timeago = "^1.0.15"
Flask = "^1.1.2"
"discord.py" = "^1.6.0"
hentai = "^3.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
